FF7: Enable custom eye texture support
Fixes #589

+++ src/ff7_opengl.cpp
@@ -280,7 +280,7 @@ void ff7_init_hooks(struct game_obj *_game_object)
 	// red XIII eye blinking
 	// #####################
 	byte ff7_redxiii_eye_fix[] = "\xEC\x79\x90\x00\x00\x00\x00\x00";
-	memcpy_code(ff7_externals.field_models_eye_blink_buffer + 0x58, ff7_redxiii_eye_fix, sizeof(ff7_redxiii_eye_fix) - 1);
+	memcpy_code((uint32_t)ff7_externals.field_models_eye_blink_buffer + 0x58, ff7_redxiii_eye_fix, sizeof(ff7_redxiii_eye_fix) - 1);
 
 	// #####################
 	// field vertical center
